src/requirements.txt
gunicorn
google-generativeai
python-dotenv